messaging_app/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-nested-routers>=0.93